Bot-App: now alarm can be added with 0.1% accuracy.

+++ BotApp/src/main/java/ru/bandit/cryptobot/menu/MenuTargetUp.java
@@ -40,7 +40,7 @@ public class MenuTargetUp extends AbstractMenuItem {
             UserTriggerEntity savedTrigger = triggersService.subscribe(userDTO,
                     new CurrencyPairDTO(queryParams.get(0), queryParams.get(1)),
                     "target_up",
-                    Integer.parseInt(queryParams.get(2)));
+                    Double.parseDouble(queryParams.get(2)));
 
             return String.format("Следим за превышением курса %s/%s выше порога %s.",
                     savedTrigger.getCurrencyPair().getCurrency1().getCurrencyNameUser(),
